requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
pycodestyle==2.10.0
python-decouple==3.6
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1